refactor: move other stuff
